fix: precio rebajado

+++ producto/models.py
@@ -15,7 +15,7 @@ class Producto(models.Model):
     marca = models.CharField(max_length=100)
     modelo = models.CharField(max_length=100)
     descripcion = models.TextField()
-    precio = models.DecimalField(max_digits=10, decimal_places=2)
+    precio = models.FloatField()
     imagen = models.URLField()
     unidades = models.IntegerField(default=0)
     capacidad_viajar = models.IntegerField(default=0)
